backend/requirements.txt
alembic==1.10.3
anyio==3.6.2
attrs==23.1.0
better-exceptions==0.3.3
build==0.10.0
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.3
crashtest==0.4.1
cryptography==40.0.2
distlib==0.3.6
dnspython==2.3.0
dpn-pyutils==3.0.5
dulwich==0.21.3
email-validator==2.0.0
fastapi==0.95.1
filelock==3.11.0
greenlet==2.0.2
h11==0.14.0
html5lib==1.1
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
importlib-metadata==6.4.1
installer==0.7.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.1.2
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
Mako==1.2.4
MarkupSafe==2.1.2
more-itertools==9.1.0
msgpack==1.0.5
orjson==3.8.10
packaging==23.1
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==2.6.2
poetry==1.4.2
poetry-core==1.5.2
poetry-plugin-export==1.3.0
psycopg2==2.9.6
ptyprocess==0.7.0
pycparser==2.21
pydantic==1.10.7
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
rapidfuzz==2.15.1
requests==2.28.2
requests-toolbelt==0.10.1
SecretStorage==3.3.3
shellingham==1.5.0.post1
six==1.16.0
slugify==0.0.1
sniffio==1.3.0
SQLAlchemy==2.0.9
SQLAlchemy-Utils==0.41.0
sqlalchemy2-stubs==0.0.2a33
starlette==0.26.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.7
trove-classifiers==2023.3.9
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
virtualenv==20.21.0
watchfiles==0.19.0
webencodings==0.5.1
websockets==11.0.1
zipp==3.15.0
